optimized selections

+++ BoostedAnalyzer/python/Selection_cff.py
@@ -42,7 +42,7 @@ DiLeptonSelectionData = cms.PSet(
 )
 
 METSelection = cms.PSet(
-    minMET = cms.double(200),
+    minMET = cms.double(250),
     maxMET = cms.double(100000),
 )
